bug
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/review@48 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ R/electronicAppendix.R
@@ -31,8 +31,8 @@ electronicAppendix <- function(
 	append.txt <- function(x)file.rename(x,paste(sep='',x,'.txt'))
 	sapply(change,append.txt)
 	if(zip)system(paste('zip -r',zipname,tmpdir))
-	else file.copy(from=tmpdir,to=as,recursive=TRUE)
-	unlink(tmpdir)
+	else file.rename(from=tmpdir,to=as)
+	if(file.exists(tmpdir))unlink(tmpdir)
 }
